requirements.txt
aioredis==2.0.1
aiosmtplib==1.1.6
aiosqlite==0.17.0
anyio==3.4.0
asgiref==3.4.1
async-timeout==4.0.2
asyncpg==0.25.0
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
databases==0.5.4
Deprecated==1.2.13
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fakeredis==1.7.0
fastapi==0.70.0
fastapi-mail==1.0.2
greenlet==1.1.2
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.3
install==1.3.5
iso8601==0.1.16
Jinja2==3.0.3
MarkupSafe==2.0.1
ormar==0.10.24
packaging==21.3
passlib==1.7.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.6
pypika-tortoise==0.1.3
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
redis==4.0.2
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.29
sqlalchemy2-stubs==0.0.2a19
sqlmodel==0.0.6
starlette==0.16.0
tortoise-orm==0.18.0
typing-extensions==4.0.1
uvicorn==0.16.0
wrapt==1.13.3
